move Azure Blob Proxy to lib dex commons and change reference in file-sink and receiver-debatcher
