+++ local_libs/lib-dex-commons/build.gradle.kts
@@ -9,7 +9,7 @@ plugins {
 }
 
 group = "gov.cdc.dex"
-version = "0.0.45-SNAPSHOT"
+version = "0.0.46-SNAPSHOT"
 
 repositories {
     maven {
